refactor: input 모델 사용으로 변경
