--- src/main/kotlin/step2/domain/operation/AdditionOperationCommand.kt
@@ -1,5 +1,0 @@
-package step2.domain.operation
-
-class AdditionOperationCommand : OperationCommand {
-    override fun operate(first: Int, second: Int) = first + second
-}
