--- src/main/kotlin/step2/domain/operation/DefaultOperationCommand.kt
@@ -1,5 +1,0 @@
-package step2.domain.operation
-
-class DefaultOperationCommand : OperationCommand {
-    override fun operate(first: Int, second: Int) = 1
-}
