--- src/main/kotlin/step2/domain/operation/DivisionOperationCommand.kt
@@ -1,7 +1,0 @@
-package step2.domain.operation
-
-class DivisionOperationCommand : OperationCommand {
-    override fun operate(first: Int, second: Int): Int {
-        return first / second
-    }
-}
