--- src/main/kotlin/step2/domain/operation/MultiplicationOperationCommand.kt
@@ -1,5 +1,0 @@
-package step2.domain.operation
-
-class MultiplicationOperationCommand : OperationCommand {
-    override fun operate(first: Int, second: Int) = first * second
-}
