--- src/main/kotlin/step2/domain/operation/OperationCommand.kt
@@ -1,5 +1,0 @@
-package step2.domain.operation
-
-interface OperationCommand {
-    fun operate(first: Int, second: Int): Int
-}
