--- src/main/kotlin/step2/domain/operation/SubtractionOperationCommand.kt
@@ -1,5 +1,0 @@
-package step2.domain.operation
-
-class SubtractionOperationCommand : OperationCommand {
-    override fun operate(first: Int, second: Int) = first - second
-}
