+++ src/main/kotlin/step2/domain/operation/command/AdditionOperationCommand.kt
@@ -1,0 +1,7 @@
+package step2.domain.operation.command
+
+import step2.domain.operation.model.InputOperationCommand
+
+class AdditionOperationCommand : OperationCommand {
+    override fun operate(input: InputOperationCommand) = input.firstNumber + input.secondNumber
+}
