+++ src/main/kotlin/step2/domain/operation/command/DefaultOperationCommand.kt
@@ -1,0 +1,7 @@
+package step2.domain.operation.command
+
+import step2.domain.operation.model.InputOperationCommand
+
+class DefaultOperationCommand : OperationCommand {
+    override fun operate(input: InputOperationCommand) = 1
+}
