+++ src/main/kotlin/step2/domain/operation/command/OperationCommand.kt
@@ -1,0 +1,7 @@
+package step2.domain.operation.command
+
+import step2.domain.operation.model.InputOperationCommand
+
+interface OperationCommand {
+    fun operate(input: InputOperationCommand): Int
+}
