+++ src/main/kotlin/step2/domain/operation/model/InputOperationCommand.kt
@@ -1,0 +1,3 @@
+package step2.domain.operation.model
+
+data class InputOperationCommand(val firstNumber: Int, val secondNumber: Int)
